+++ packages/lib/test/parent/parent.test.ts
@@ -2,7 +2,7 @@ import { remove, set } from "mobx"
 import {
   detach,
   findParent,
-  findParentWithPath,
+  findParentPath,
   fromSnapshot,
   getChildrenObjects,
   getParentPath,
@@ -69,28 +69,28 @@ test("parent", () => {
 
   expect(findParent(p.p2!, parent => parent instanceof P2)).toBe(undefined)
 
-  expect(findParentWithPath(p.p2!, parent => parent instanceof P)).toEqual({
+  expect(findParentPath(p.p2!, parent => parent instanceof P)).toEqual({
     parent: p,
     path: ["p2"],
   })
-  expect(findParentWithPath(p.p2!, parent => parent instanceof P, 0)).toEqual({
+  expect(findParentPath(p.p2!, parent => parent instanceof P, 0)).toEqual({
     parent: p,
     path: ["p2"],
   })
-  expect(findParentWithPath(p.p2!, parent => parent instanceof P, 1)).toEqual({
+  expect(findParentPath(p.p2!, parent => parent instanceof P, 1)).toEqual({
     parent: p,
     path: ["p2"],
   })
-  expect(findParentWithPath(p.p2!, parent => parent instanceof P, 2)).toEqual({
+  expect(findParentPath(p.p2!, parent => parent instanceof P, 2)).toEqual({
     parent: p,
     path: ["p2"],
   })
-  expect(findParentWithPath(p.p2!, parent => parent instanceof P, 3)).toEqual({
+  expect(findParentPath(p.p2!, parent => parent instanceof P, 3)).toEqual({
     parent: p,
     path: ["p2"],
   })
 
-  expect(findParentWithPath(p.p2!, parent => parent instanceof P2)).toBe(undefined)
+  expect(findParentPath(p.p2!, parent => parent instanceof P2)).toBe(undefined)
 
   expect(getParentPath(p)).toBeUndefined()
   expect(() => getParentPath(p.$)).toThrow($errorMessage)